// server/src/logging/enums.ts
export enum LogLevel {
  Panic = "panic", // 0
  Alert = "alert", // 1
  Critical = "crit", // 2
  Error = "error", // 3
  Warn = "warn", // 4
  Notice = "notice", // 5
  Info = "info", // 6
  Debug = "debug", // 7
}

export enum LogType {
  Error = "error",
  Request = "request",
  Response = "response",
}